ZKSyncLite code organization: No longer an ethereum submodule
Removed it under modules/l2 and made it in its own chain manager
